+++ src/next/data/history.ts
@@ -1,5 +1,5 @@
 import * as vscode from "vscode"
-import { RCSNext } from ".."
+import { RocketChatStatus } from ".."
 
 const statusHistoryLimitField = "statusHistoryLimit"
 
@@ -7,9 +7,9 @@ export async function deleteHistory(ctx: vscode.ExtensionContext): Promise<void>
     await ctx.globalState.update(statusHistoryLimitField, undefined)
 }
 
-export async function getHistory(ctx: vscode.ExtensionContext): Promise<RCSNext.Base.StoredStatus[]> {
+export async function getHistory(ctx: vscode.ExtensionContext): Promise<RocketChatStatus.Base.StoredStatus[]> {
     const history = ctx.globalState
-        .get<RCSNext.Base.StoredStatus[]>(statusHistoryLimitField)
+        .get<RocketChatStatus.Base.StoredStatus[]>(statusHistoryLimitField)
 
     if (history === undefined) {
         return []
@@ -18,6 +18,6 @@ export async function getHistory(ctx: vscode.ExtensionContext): Promise<RCSNext.
     return history
 }
 
-export async function setHistory(ctx: vscode.ExtensionContext, history: RCSNext.Base.StoredStatus[]): Promise<void> {
+export async function setHistory(ctx: vscode.ExtensionContext, history: RocketChatStatus.Base.StoredStatus[]): Promise<void> {
     await ctx.globalState.update(statusHistoryLimitField, history)
 }