+++ src/next/data/setup.ts
@@ -1,5 +1,5 @@
 import * as vscode from "vscode"
-import { Maybe, RCSNext } from ".."
+import { Maybe, RocketChatStatus } from ".."
 
 export const setupField = "setup"
 
@@ -7,7 +7,7 @@ export async function deleteSetup(ctx: vscode.ExtensionContext): Promise<void> {
     await ctx.secrets.delete(setupField)
 }
 
-export async function getSetup(ctx: vscode.ExtensionContext): Promise<Maybe<RCSNext.Base.Setup>> {
+export async function getSetup(ctx: vscode.ExtensionContext): Promise<Maybe<RocketChatStatus.Base.Setup>> {
     const rawSetup = await ctx.secrets.get(setupField)
 
     if (rawSetup === undefined) {
@@ -23,7 +23,7 @@ export async function getSetup(ctx: vscode.ExtensionContext): Promise<Maybe<RCSN
             "baseUrl" in parsed &&
             "userId" in parsed
         ) {
-            return parsed as RCSNext.Base.Setup
+            return parsed as RocketChatStatus.Base.Setup
         }
 
         return undefined
@@ -33,6 +33,6 @@ export async function getSetup(ctx: vscode.ExtensionContext): Promise<Maybe<RCSN
 
 }
 
-export async function setSetup(ctx: vscode.ExtensionContext, setup: RCSNext.Base.Setup): Promise<void> {
+export async function setSetup(ctx: vscode.ExtensionContext, setup: RocketChatStatus.Base.Setup): Promise<void> {
     await ctx.secrets.store(setupField, JSON.stringify(setup))
 }